implement more functions

+++ include/string.h
@@ -6,7 +6,7 @@
 
 size_t strlen (const char * str);
 char * strcpy (char * restrict d, const char * restrict s);
-void * memset (void * dest, uint32_t c, uint64_t n);
+void * memset (void * dest, uint32_t c, size_t n);
 void * memcpy (void * restrict dest, const void * restrict src, size_t n);
 void * memmove (void * dest, const void * src, size_t n);
 int memcmp (const void * vl, const void * vr, size_t n);
