+++ src/string/memset.c
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <stdint.h>
 
-void * memset (void * dest, uint32_t c, uint64_t n) {
+void * memset (void * dest, uint32_t c, size_t n) {
   unsigned char * s = dest;
   uint64_t k;
